+++ examples/resources/meraki_network_wireless_radio_rrm/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_network_wireless_radio_rrm.example "<network_id>"